Rename and separate to WrodFrequencyFilter
