+++ chariot/transformer/token/__init__.py
@@ -1,3 +1,4 @@
-from .stopword_filter import StopwordFilter, FrequencyBasedStopwordFilter
+from .stopword_filter import StopwordFilter
 from .base_form_normalizer import BaseFormNormalizer
 from .number_normalizer import NumberNormalizer
+from .word_frequency_filter import WordFrequencyFilter
